Don't use `cache-max-size-lazy-result` setting for the root operation, even when it's lazy (#1709)
Since https://github.com/ad-freiburg/qlever/pull/1638, QLever has the runtime parameter `lazy-result-max-cache-size`, with a deliberately small default value of 5 MB. However, the final result (of the root operation) should be cached not depending on that size, but depending on the regular `cache-max-size`, `cache-max-size-single-entry`, and `cache-max-num-entries`, even if the last operation is (output-)lazy.  This is now indeed so. Fixes #1701 (which provides a good example of a query, where the old behavior is weird and the new behavior makes much more sense).

Use the occasion to give the runtime parameter a more consistent name, namely `cache-max-size-lazy-result`.