Better error message on parallel turtle parsing of scattered base and prefix declarations (#1807)
The parallel Turtle parser requires that all `PREFIX` and `BASE` declarations stand in a single block at the beginning of the file. 
Until now, prefix declarations after this initial block led to an exception about the prefix being undeclared, which was highly misleading. With this PR, this case triggers a proper exception about PREFIX declarations being illegal in the middle of a Turtle file with the parallel parser, together with some guidance to mitigate this issue by either refactoring the input, or deactivating the parallel parser.

Fixes #1794 by providing a better message for the described error.