Fix the Conan MacOS build (#1814)
For some time now, the CI build on MacOS using the Conan package manager has been broken, because the building of the ICU dependency failed.  This PR fixes this build again.

The background is, that we have to link against a custom (brew installed) standard library, which makes the whole build process a lot more brittle as several flags have to be set manually.