Truncate echo of SPARQL operation strings (#1846)
The raw operation strings (queries or updates) are echoed in the response for `application/qlever-results+json` and the console during execution. These operation strings can become quite long, especially for SPARQL UPDATE requests that contain a lot of explicit triples. This PR truncates the operation strings to a fixed length. The error location is only sent if it is completely within the string (not truncated).