Better transitive path planning (#1824)
Change the query planning in a following way:
A Join between a UNION and another operation is distributed if one of the children of the union (possibly recursively) is (or recursively contains in another UNION) an unbound property path.
This returns much better query plans in particular for complex transitive paths such as `?x <a>/(<b>* | <c>*) ?y`.