Implement `Union` that can keep sort order (#1864)
If the result of a `UNION` has to be sorted, then the sort is propagated to the children of the UNION and the UNION is performed by a binary merge operation of the already sorted inputs.
This is especially beneficial when some of the inputs of the UNION are already implicitly sorted (e.g. because they are IndexScans).
Note: The query planner can currently not yet use the full potential of this feature, as the decision on the sorting of the subtrees of the UNION is made before we know that the result of the UNION has to be sorted, but at least for IndexScans this can be easily fixed.