Fix `getResultSizeImpl` for small scans with updates (#1878)
There was a bug in `CompressedRelationReader::getResultSizeImpl` that ignored updates for index scans with only one or two blocks. This was especially dramatic for the exact result size which could lead to whole queries being optimized out because the exact size of an index scan was falsely reported as zero (because a block contained no triples in the initial index, but contained updates).

Fixes #1861