Let `DISTINCT` sort its subtree (#1854)
So far, the `Distinct` class had the invariant that its subtree was already sorted appropriately. With this PR, this invariant is enforced by the constructor of the `Distinct` class by calling `createSortedTree`, which makes the behavior more consistent with other operations and simplifies code in the query planner. 
Additionally, a variation of  `createSortedTree` was implemented that makes sure that `Distinct` doesn't resort the subtree if the the subtree is sorted by a permutation of the desired sort columns, as the used algorithm also works in this case.