Fix assertion error in `LocalVocab` when a local vocab is merged with a clone of itself. (#1899)
There was an assertion failure when a local vocab was first cloned and then merged with itself. This pattern seems odd but is sometimes necessary or can happen when lazy and materialized operations are alternating or when the cache is involved.
This PR fixes this assertion failure by also deduplication wrt the `primary word set` of the local vocab when merging.

Also implement an overload for `LocalVocab::mergeWith` that simply takes a single other vocab without having to arbitrary construct a range.

Fixes #1890