Allow REGEX expressions with a dynamically computed regular expression (#1889)
Add support for the `REGEX` function in the case where the regular expression is dynamically computed by the query, e.g.
`REGEX(?input, ?regex)` or `REGEX(?input, CONCAT("bla", "*"))`. Previously, QLever only supported hardcoded regular expressions like `REGEX(?input, "bla*")`.
The implementation optimizes for the case, that the regex is stored in a variable with a low multiplicity, so for example for the case `REGEX(?input, ?regex)`, where the `?regex` only binds to a few distinct values by using a simple LRU cache in the `RegexValueGetter`.
Also clean up the code for the `Regex`, `Replace` and `Prefix` expressions.
Fixes #1821