Fix cache key of `UNION` operation (#1933)
So far, QLever's cache keys do not take the set of selected variables into account. This has the advantage that results can be reused when only the set of selected variables changes. However, it can lead to nasty bugs with `UNION` when the result for one side is found in the cache, but for a different set of selected variables. The following sequence demonstrates the problem (before this change, the result for the second query was taken from the cache, which is wrong).

https://qlever.cs.uni-freiburg.de/api/wikidata?cmd=clear-cache
https://qlever.cs.uni-freiburg.de/wikidata/fwYscF?exec=true
https://qlever.cs.uni-freiburg.de/wikidata/TV20Jn?exec=true

This is now fixed by adding the sequence of selected variables for each side to the cache key of a `UNION` operation. Fixes #1900 and potentially also #1898.