Fix subtle bug in query planner (involving plans without variables) (#1955)
There was a bug in the query planner when two query plan alternatives have no variables in common, as it happens for the query below (which doesn't make sense except that it manifested the bug). The query planner should then pick the cheaper of the two alternatives and discard the other. But so far, the two alternatives were erroneously joined via a Cartesian product. This is now fixed.

```sparql
SELECT * WHERE {
  1 a* 1
}
```