Add strong type `KeyOrder` for the key order of a permutation (#1954)
So far, the type of this order was hard-coded as `std::array<size_t, 3>` in many places. Now, there is an own type `KeyOrder`, which stores an `std::array<uint8_t, 4>`. For example, `{1, 0, 2, 3}` represents the permutation `PSOG`.

On the side, improve the `IdTriple` class (which stores a quad with `N` additional payload `Id`s) such that for `N = 0` (no additional payload) no additional space is used.