Unify interface of the different vocabulary `WordWriter`s (#1979)
The `WordWriter` types of all vocabulary implementations now return the index of an inserted word, support an argument `bool isExternal` (which mostly is a dummy)  and have a method `readableName()` to get/set a name. This unifies their interface such that we can easily exchange the different implementations.

The vocabulary merging now directly uses the index of a word according to the vocabulary. So far, it computed that index itself, which was redundant and error-prone.