Make spatial join with `<libspatialjoin>` cancellable (#1953)
Spatial joins with `<libspatialjoin>` were not cancellable so far. This is fixed with this PR. Cancellation is checked after each parsed `WKT` string, after every 10,000th step during sweeping, and during each iteration of the result aggregation from the `libspatialjoin` threads. To allow checks within the sweep process of `libspatialjoin`, the `SpatialjoinConfig` now accepts a callback function which is called every 10,000th step.

During sweeping in `libspatialjoin`, an atomic bool `_canceled` is set to true as soon as the callback function throws. This causes all threads to simply discard any job they retrieve from the job queue, so the job queue quickly runs empty. After all threads have `join()`ed, the exception from the callback function is rethrown, enabling QLever to detect it.

A similar mechanism is used to gracefully spin down all parsing threads if the `WKTParser` object is deleted before the `done()` method is called on it (which is what happens if the `cancellationWrapper` throws).