Allow `SERVICE` result to be cached if explicitly so desired (#1981)
With #1874, the results of `SERVICE` operations are no longer cached for two reasons: in order to guarantee correctness (in case the `SERVICE` endpoint is update between requests) and to enable the "VALUES" optimization described in #1874 (which can reduce the time for processing the operation significantly when the result from the `SERVICE` operation is large but is strongly constrained by the rest of the query).

However, in certain scenarios (in particular, when the `SERVICE` endpoint is under one's own control), caching the result of a `SERVICE` request can be desirable. This PR therefore introduces the command line flag `--cache-service-results`, which is off by default, that reverts to the old behavior before #1874. That is, with this flag the result of a `SERVICE` operations is cached just like the results of other operations, but there is no "VALUES" optimization when computing the result for the first time.

Fixes #1929