Fix two more corner cases of property paths with zero length (#1963)
For a query like `SELECT * { ?x a* ... }` the result should be empty if `...` does not occur as a subject or object in any triple, and the result should be `...` if it does. However, the current behavior is not correct for `...` that do not occur as a subject or object in any triple and when one of the following two holds: (1) `...` is a value that can be folded into the `Id` (then the result is `...` but should be empty) or (2) `...` is a literal or IRI that does not exist in the original dataset (then an exception is thrown, even if the literal or IRI was added in an update operation). These corner cases are now fixed.

On the side, queries like `SELECT * { 1 a* 2 }` are now rewritten to the equivalent `SELECT * { 1 a+ 2 }`, which is much more efficient to compute (if the subject and object are fixed but different, there is no match for the "empty path").

The result for queries like `SELECT * { VALUES ?x { 1 } ?x a* ?x }` (one of the cases of the W3C test suite) is still incorrect. This will be fixed in one the next PRs.