Give `ProgressBar` users control over timer if needed (#2006)
So far, the timer for a `ProgressBar` was always started as soon as the `ProgressBar` was created. However, some input generation processes (for example, `osm2rdf` feeding OSM triples to QLever's index builder) have a long startup time before they produce the first item but then produce items fast. In such a scenario. if we count that startup time towards the first batch of items, the average batch processing time reported by the `ProgressBar` is heavily distorted.

The `ProgressBar` class now has a `getTimer()` method that gives the user more control. This is used in `IndexImpl::passFileForVocabulary()` to start timing only once the first triple has been parsed.