Better error message when running out of memory (#2057)
So far, QLever returned an error message of the following kind when `MEMORY_FOR_QUERIES` was exceeded:

`"Tried to allocate ..., but only ... were available. Clear the cache or allow more memory for QLever during startup"`

The second sentence of this message was very misleading because QLever does that automatically already. That is, when you get this error message, there is simply not enough memory for processing the query according to the query plan determined by QLever.

The first sentence is slightly misleading because the amounts named are more of an implementation detail. In particular, the first amount is not the whole memory needed to process the respective query, but just the memory needed for one of usually multiple allocations (the one that failed).

We now drop the second sentence, but leave the first. Resolves #2020