Allow each `Operation` to specify how to make sure its result is sorted (#1946)
Introduce an interface that allows every `Operation` to specify how to make sure that its result is sorted according to a given ordering of the columns. Operations which cannot do that themselves return `std::nullopt`, which causes the operation to be wrapped inside a `Sort` operation. However, if the operation has a more efficient way to achieve the desired result, it can return an alternative execution tree instead.