Implement `Load` operation, as a preparation for `LOAD` (#2050)
Implement a `Load` operation (subclass of `Operation`), which reads RDF Turtle data from a given IRI and returns the corresponding triples as a `Result` table with three columns `?s`, `?p`, `?o`. This is preparation for the SPARQL update operation `LOAD`, see https://www.w3.org/TR/sparql11-update/#load, which reads RDF Turtle data from a given IRI and inserts it into a given graph.

NOTE: This currently uses the `RdfStringParser`, which is not particularly efficient (30s / 1M triples). Also, the operation is neither input- nor output-lazy. That is, the RDF Turtle data is read fully into RAM and the `Result` table is fully materialized. This is work for future PRs.