Make `LocatedTriplesPerBlock::numTriples` more efficient (#2083)
So far, `LocatedTriplesPerBlock::numTriples` iterated over the `LocatedTriplesPerBlock` for each block, counted the number of triples with `insertOrDelete_ == true` and `insertOrDelete_ == false`, and returned that pair. This turned to be very inefficient because the located triples per block are stored in a `std::set` and iterating over an `std::set` is terrifyingly slow. As a consequence, many queries involving update triples were also slow because the size estimator for index scans uses `LocatedTriplesPerBlock::numTriples`.  This is now fixed by simply returning the number of triples in the block as an upper bound for both the number of inserted and the number of deleted triples in the block. Fixes #2072, which explains all the details behind this.

On the side, improve a variable name (`insertOrDelete` was called `shouldExist` before) and several comments. Also update some of the copyright headers to the latest format. The actual code change is very small, but its consequences are huge.

With this change, the first ten example queries for Wikidata are now only a few percent slower when running with the full set of update triples (a bit over 100 M triples) needed to bring the SPARQL endpoint (that was built from the weekly dump) in sync with the current version.