Preparation integration of  lightweight `BlockMetadataRanges` into `IndexScan` (#1958)
So far, we prefilter the `CompressedBlockMetadata` associated with an `IndexScan` twice, once during query planning for the size estimate, and then again when processing the query. In both cases the prefiltered ``CompressedBlockMetadata` are returned as a materialized vector. This slows down queries that involve large index scans significantly, for example:
```
SELECT * WHERE {
  ?s ?p ?o 
  FILTER (?o > <h>)
}
LIMIT 1000
```
This change implements the environment to perform the materialization of the prefiltered `BlockMetadataRanges` when performing a `lazyScan`. Specifically, it prepares those methods in `IndexScan` that are relevant for `CompressedRelation.h` to take `ScanSpecAndBlocks` as an argument, where `ScanSpecAndBlocks` holds the newly introduced lightweight  `BlockMetadataRanges`, which represents the relevant (in the future prefiltered) subranges of the underlying index.

NOTE: This is the first of two changes. The second change will be to add the `ScanSpecAndBlocks` as a variable to the `IndexScan` class, while directly adding the prefiltered `BlockMetadataRanges` during its construction.