Make includes consistent regarding the use of `"..."` and `<...>` (#2097)
QLever's rule for includes is that `#include <...>` is used for system headers and `#include "..."` is used for QLever headers. There were a few places, were this was not done consistently, and these are now fixed. While at it, remove some unused includes and add some missing includes (that only worked implicitly via other includes).