Spatial join service now accepts floating point values for `<maxDistance>` (#2093)
The `<maxDistance>` parameter of the spatial join service now accepts a floating point value and is also stored as a floating point value. In particular, the distance, which is measured in meters, can now be smaller than one meter. This follows the suggestion from <https://github.com/ad-freiburg/qlever/pull/1935#discussion_r2120272456>.