Implement `LOAD` (#2089)
Implement the Graph Update operation `LOAD`.

On the side, fix a lifetime issue with the `LocalVocab` of the `Result` object produced by an update query. Namely, the `Result` only lived inside `computeGraphUpdateQuads`, so that the `LocalVocab` IDs were no longer live when being passed to `DeltaTriples::insertTriples` or `DeltaTriples::deleteTriples`.