Fix behavior of `*` of `?` when subject or object is bound (#1999)
Fix the behavior of property paths with `*` or `?` when the subject or object is bound. For example, the following query should have a non-empty result if and only if the literal `1` is present in the graph. So far, the query produces all values of `?x` as result. The current fix adds an extra join operation. This join will be made more efficient once a subset of the changes proposed in #1666 are committed.
```sparql
SELECT * {
  VALUES ?x { 1 }
  ?x a* ?x
}
```