Add `SplitVocabulary` class, as a preparation for a separate geo vocabulary (#2111)
Add a class `SplitVocabulary`, which can store words in different vocabularies based on a given split function. It uses marker bits in the vocabulary index to specify which vocabulary a word belongs to. This is preparation for #1983, where the `SplitVocabulary` is used to store WKT literals in their own vocabulary.