Fix `CompressedRelation::getFirstAndLastBlock` when there are updates (#2120)
So far, `CompressedRelation::getFirstAndLastBlock` read the first and last block of the given range `metadataAndBlocks` in order to determine the first and last triple. However, when there are updates, all the triples in a block might be deleted, so we have to find the first non-empty block from the beginning and from the end of the range. In the typical case, when both the first and the last block of the given range are non-empty, the time complexity of the function does not increase.

NOTE: For large relations, the reading of the first and last block could be saved by looking up the respective information in the (augmented) metadata. However, we don't expect a significant performance gain from that + we never did that so far, also before we had updates; see the discussion at the end of the PR.

Co-authored-by: Hannah Bast <bast@cs.uni-freiburg.de>